+++ recipes/oozie_client.rb
@@ -1,0 +1,7 @@
+#
+# Cookbook Name:: mapr_installation
+# Recipe:: oozie_client
+#
+# Copyright (c) 2016 Synchronoss Technologies, Inc., All Rights Reserved.
+
+include_recipe 'mapr_installation::_oozie_common'